refactor(v0.52 phase 4): server/ + channels/ 신설, gateway/ 폐기
매핑 매트릭스 §3.2 적용. core/gateway/ 잡탕 디렉토리를 process binding
명시적인 두 개로 분해.

이름 정당화:
- server/        — daemon process 호스팅 인프라 (IPC poller, supervised tasks)
- channels/      — 외부 IO surface 추상화 (Slack/Discord/Telegram). first-class
                   capability (Hermes/OpenClaw 패턴)

이동:
- core/gateway/pollers/cli_poller.py     → core/server/ipc_server/poller.py
- core/gateway/pollers/base.py           → core/server/supervised/poller_base.py
- core/gateway/pollers/{slack,discord,telegram}_poller.py
                                         → core/server/supervised/<same>
- core/gateway/channel_manager.py        → core/channels/binding.py
- core/gateway/models.py                 → core/channels/models.py
- core/gateway/slack_formatter.py        → core/channels/slack_formatter.py
- core/gateway/webhook_handler.py        → core/server/supervised/webhook_handler.py
- core/gateway/shared_services.py        → core/server/supervised/services.py

폐기: core/gateway/ (빈 디렉토리)

Import 경로 갱신: 14 파일 (sed 자동)

Invariant test 갱신:
- DAEMON_DIRS 에서 core/gateway/ 제거, core/server/ + core/channels/ 추가
- daemon-side native I/O 검사가 새 디렉토리에 자동 적용

검증:
- ruff + format clean (3 import 자동수정)
- mypy strict clean (230 modules, +2 신설)
- pytest 4081 passed (1 skipped, 19 deselected)
- E2E `geode analyze "Cowboy Bebop" --dry-run` → A (68.4) unchanged

다음 phase: automation/ → scheduler/ + config.py → config/ 디렉토리화
