+++ core/channels/binding.py
@@ -11,9 +11,9 @@ import threading
 from collections.abc import Callable
 from typing import Any
 
-from core.gateway.models import ChannelBinding, InboundMessage
-from core.gateway.pollers.base import BasePoller
+from core.channels.models import ChannelBinding, InboundMessage
 from core.memory.session_key import build_gateway_session_key
+from core.server.supervised.poller_base import BasePoller
 
 MessageProcessor = Callable[[str, dict[str, Any]], str]
