+++ core/channels/slack_formatter.py
