--- core/gateway/__init__.py
@@ -1,5 +1,0 @@
-"""Gateway — inbound message routing from external messaging services.
-
-OpenClaw Gateway pattern: ChannelBinding-based message routing
-with per-channel pollers (Slack, Discord, Telegram).
-"""
