+++ tests/test_gateway.py
@@ -8,12 +8,12 @@ from __future__ import annotations
 import time
 from unittest.mock import MagicMock, patch
 
-from core.gateway.channel_manager import ChannelManager, get_gateway, set_gateway
-from core.gateway.models import ChannelBinding, InboundMessage
-from core.gateway.pollers.base import BasePoller
-from core.gateway.pollers.discord_poller import DiscordPoller
-from core.gateway.pollers.slack_poller import SlackPoller
-from core.gateway.pollers.telegram_poller import TelegramPoller
+from core.channels.binding import ChannelManager, get_gateway, set_gateway
+from core.channels.models import ChannelBinding, InboundMessage
+from core.server.supervised.discord_poller import DiscordPoller
+from core.server.supervised.poller_base import BasePoller
+from core.server.supervised.slack_poller import SlackPoller
+from core.server.supervised.telegram_poller import TelegramPoller
 
 # ---------------------------------------------------------------------------
 # InboundMessage
