+++ tests/test_slack_formatter.py
@@ -1,8 +1,8 @@
-"""Tests for core.gateway.slack_formatter — Markdown → Slack mrkdwn conversion."""
+"""Tests for core.channels.slack_formatter — Markdown → Slack mrkdwn conversion."""
 
 from __future__ import annotations
 
-from core.gateway.slack_formatter import markdown_to_slack_mrkdwn
+from core.channels.slack_formatter import markdown_to_slack_mrkdwn
 
 _ZWS = "\u200b"
